Added Model, Data Contysext and a Database.
Added a Model directory to our system. It contains one class called "Values" in it. Also created a Data Context using this same class and used Entity Framework and Sqlite to create a database capable of storing such values. Used DB Browser to add sample data into the databse and modified the Values Controller to ensure these data can be retrieved and viewd in the requests using Postman for now. Also change functions associated with the routes to include an "I" prefix so that we return a status code along with the content.