Added Security Measures
Added some level of security to our application by setting up authorization and authentication features with SHA512 encryption and Token Generation and usability. At this stage, the application allows for the table of values to be retrieved only through proper authentication and by addion of token into our header with our GET request.